Implement exponential and randomized strategies
We first tried to implement Google's exponential backoff strategy as one
fn, but after a bit of thinking it turns out that a
multiplicative (exponential) strategy is a separate concern from
randomizing delays in order to avoid overwhelming a remote service when
it comes back up. So we split the initial exponential-strategy into two
fns: exponential-strategy generates an inifinite multiplicative strategy
and randomize-strategy applies a random factor to an existing
strategy. Now randomize-strategy can be used with any of the existing
strategies!