Remove CLJX stuff
I can't see a way (at the moment) for this being useful in CLJS, because
of the sleep in the retry loop. It would make sense if the whole thing
was run in a go loop with core.async, but I haven't had enough time to
think it through. So maybe some time in the future.