Remove the defsleeplesstest macro
I don't think that the amount of duplication if worth adding a macro at this
stage, although I might change my mind later. I also don't like the current
name, which I feel is long and hard to read. I'd almost be tempted to just
overwrite `deftest`, or define a `defagaintest` (which also is hard to read) and
use that macro for all `again` tests…